serial.c Rx String

+++ 1_Standart_Projekte/Atmega1284p_std/serial.c
@@ -70,7 +70,7 @@ unsigned char USART_Receive_STRING(unsigned char *st)
 	}
 	
 	st[digit] = '\0';			//add NULL termination
-	return st;
+	return (int)st;
 }
 
 //============= SPI =======================
